+++ bench/tests/test_routing.py
@@ -51,13 +51,12 @@ def test_op(n_tokens, n_expts_tot, n_expts_act, block_m):
     ref_logits = tri_logits.clone()
     ref_routing_data, ref_gather, ref_scatter = routing_torch(ref_logits, n_expts_act)
     tri_routing_data, tri_gather, tri_scatter = routing(tri_logits, n_expts_act)
+    ref_metadata = ref_expt_data(ref_routing_data, n_tokens * n_expts_act, block_m)
+    tri_metadata = tri_routing_data.expt_data(n_tokens * n_expts_act, block_m).buffer
 
     assert_close(ref_routing_data.gate_scal, tri_routing_data.gate_scal, 2e-2, 4e-3)
     assert_equal(ref_routing_data.expt_hist, tri_routing_data.expt_hist)
-    # assert_equal(
-    #     ref_expt_data(ref_routing_data, n_tokens * n_expts_act, block_m),
-    #     tri_routing_data.expt_data(n_tokens * n_expts_act, block_m).buffer,
-    # )
+    assert_equal(ref_metadata, tri_metadata)
     assert ref_routing_data.n_expts_tot == ref_routing_data.n_expts_tot
     assert ref_routing_data.n_expts_act == ref_routing_data.n_expts_act
