+++ bench/triton_bench/routing_details/__init__.py
@@ -1,0 +1,3 @@
+from . import histogram
+from . import bitmatrix
+from . import indexing
